+++ gear/form_qc_coordinator/src/python/form_qc_coordinator_app/visits.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Dict, List, Optional
+from typing import Any, Dict, List, Literal, Optional
 
 from configs.ingest_configs import FormProjectConfigs, ModuleConfigs
 from flywheel_adaptor.flywheel_proxy import FlywheelProxy
@@ -52,6 +52,8 @@ class VisitsLookupHelper:
         module_configs: ModuleConfigs,
         cutoff_date: Optional[str] = None,
         search_op: Optional[str] = ">=",
+        missing_data_strategy: Literal["drop-row", "none"] = "drop-row",
+        add_timestamp: bool = False,
     ) -> Optional[List[Dict[str, str]]]:
         """Get the list of visits for this participant for the specified
         module. If cutoff_date specified, get the visits having a visit date on
@@ -65,6 +67,8 @@ class VisitsLookupHelper:
             module_configs: form ingest configs for the module
             cutoff_date (optional): If specified, filter visits on cutoff_date
             search_op (optional): Operator to filter visit date, defaults to >=
+            missing_data_strategy: missing_data_strategy, default 'drop-row'
+            add_timestamp: whether to add validation timestamp to result, default False
 
         Returns:
             List[Dict]: List of visits matching with the specified cutoff date
@@ -73,9 +77,11 @@ class VisitsLookupHelper:
         title = f"{module} visits for participant {self.__subject.label}"
 
         ptid_key = MetadataKeys.get_column_key(FieldNames.PTID)
+        naccid_key = MetadataKeys.get_column_key(FieldNames.NACCID)
         date_col_key = MetadataKeys.get_column_key(module_configs.date_field)
-        columns = [
+        columns: List[Any] = [
             ptid_key,
+            naccid_key,
             date_col_key,
             "file.name",
             "file.file_id",
@@ -85,6 +91,11 @@ class VisitsLookupHelper:
         if FieldNames.VISITNUM in module_configs.required_fields:
             visitnum_key = MetadataKeys.get_column_key(FieldNames.VISITNUM)
             columns.append(visitnum_key)
+
+        if add_timestamp:
+            timestamp_key = f"file.info.{MetadataKeys.VALIDATED_TIMESTAMP}"
+            timestamp_label = f"{module}-{MetadataKeys.VALIDATED_TIMESTAMP}"
+            columns.append((timestamp_key, timestamp_label))
 
         filters = f"acquisition.label={module}"
 
@@ -96,6 +107,7 @@ class VisitsLookupHelper:
             dv_title=title,
             columns=columns,
             filters=filters,
+            missing_data_strategy=missing_data_strategy,
         )
 
     def find_module_visits_with_matching_visitdate(
@@ -126,11 +138,13 @@ class VisitsLookupHelper:
         title = f"{module} visits for participant {self.__subject.label}"
 
         ptid_key = MetadataKeys.get_column_key(FieldNames.PTID)
+        naccid_key = MetadataKeys.get_column_key(FieldNames.NACCID)
         date_col_key = MetadataKeys.get_column_key(module_configs.date_field)
         timestamp_key = f"file.info.{MetadataKeys.VALIDATED_TIMESTAMP}"
         timestamp_label = f"{module}-{MetadataKeys.VALIDATED_TIMESTAMP}"
         columns = [
             ptid_key,
+            naccid_key,
             date_col_key,
             "file.name",
             "file.file_id",